Move conn.query.select() to conn.select()
