+++ wikia/common/mw_database/connection.py
@@ -6,17 +6,17 @@ import time
 import six
 import sys
 
+from .query_builder import SqlBuilderMixin
 import wikia.common.logger
 
 
 logger = logging.getLogger(__name__)
 
 
-class Connection(object):
+class Connection(SqlBuilderMixin):
     def __init__(self, raw_connection, connection_info=None):
         self.raw_connection = raw_connection
         self.connection_info = connection_info
-        self.query = ConnectionQueryBuilder(self)
         self.__logger = None
 
     @property
@@ -34,7 +34,7 @@ class Connection(object):
     def cursor(self):
         return self.raw_connection.cursor()
 
-    def _query(self, query, *args, **kwargs):
+    def query(self, query, *args, **kwargs):
         log_text = 'SQL Query: {}'.format(query)
         if 'args' in kwargs and len(kwargs['args']) > 0:
             if "\n" in query:
@@ -112,145 +112,6 @@ class Connection(object):
 
     def commit(self):
         self.raw_connection.commit()
-
-
-
-class SqlBuilderMixin(object):
-    def query(self, *args, **kwargs):
-        raise NotImplementedError('Inheritors must overrie SqlBuilderMixin.query')
-
-    def select_as_dicts(self, table, what, where):
-        return self.select(table, what, where).to_dicts
-
-    def select(self, table, what, where):
-        """
-        Execute SELECT statement
-        :param table:
-        :param what:
-        :param where:
-        :return:
-        :rtype: QueryResult
-        """
-        sql_data = {}
-        where_clause = self.where(where, sql_data)
-        sql = 'SELECT {} FROM {} WHERE {};'.format(what, table, where_clause)
-
-        return self.query(sql, args=sql_data)
-
-    def insert(self, table, data, ignore_errors=False):
-        """
-        Execute INSERT statement
-
-        :param table: Table name
-        :param data: Dictionary with data
-        :param ignore_errors: Option flag to ignore duplicated-class errors durign query execution
-        :return:
-        :rtype: QueryResult
-        """
-        columns = []
-        values = []
-        sql_data = {}
-        for column, value in data.items():
-            columns.append(column)
-            sql_value, is_value = self.add_value(value, sql_data, column)
-            if not is_value:
-                raise ValueError('insert accepts only value literals')
-            values.append(sql_value)
-        ignore = ''
-        if ignore_errors:
-            ignore = 'IGNORE '
-
-        sql = 'INSERT {}INTO {}({}) VALUES ({});'.format(ignore, table, ', '.join(columns), ', '.join(values))
-
-        return self.query(sql, args=sql_data)
-
-    def update(self, table, data, conds):
-        """
-        Execute UPDATE statement
-
-        :param table: Table name
-        :param data: Dictionary with data to be updated
-        :param conds: Conditions
-        :return:
-        :rtype: QueryResult
-        """
-        sql_data = {}
-        set_clause = []
-        where_clause = []
-        for k, v in data.items():
-            self.add_condition(k, v, set_clause, sql_data, 'data_')
-        where_clause = self.where(conds, sql_data, 'conds_')
-
-        sql = 'UPDATE {} SET {} WHERE {};'.format(table, ', '.join(set_clause), where_clause)
-
-        return self.query(sql, args=sql_data)
-
-    def delete(self, table, conds):
-        """
-        Execute DELETE statement
-
-        :param table: Table name
-        :param conds: Conditions
-        :return:
-        :rtype: QueryResult
-        """
-        sql_data = {}
-        sql = 'DELETE FROM {} WHERE {};'.format(table, self.where(conds, sql_data))
-
-        return self.query(sql, args=sql_data)
-
-    def where(self, conds, sql_data, prefix=''):
-        clause = []
-        if conds is not None:
-            for k, v in conds.items():
-                self.add_condition(k, v, clause, sql_data, prefix)
-
-        if len(clause) == 0:
-            clause = ['1 = 1']
-
-        return ' AND '.join(clause)
-
-    def add_condition(self, key, value, clause, sql_data, prefix=''):
-        sql_value, is_value = self.add_value(value, sql_data, prefix + key)
-        if is_value:
-            clause.append('{} = {}'.format(key, sql_value))
-        else:
-            clause.append(sql_value)
-
-    def add_value(self, value, sql_data, value_name):
-        is_sql_literal = hasattr(value, 'IS_SQL_LITERAL')
-
-        if not is_sql_literal:
-            sql_data[value_name] = value
-            return '%({})s'.format(value_name), True
-        else:
-            sql_data.update(value.args)
-            return value.text, value.is_value
-
-
-class ConnectionQueryBuilder(SqlBuilderMixin):
-    def __init__(self, connection):
-        self.connection = connection
-
-    def __call__(self, *args, **kwargs):
-        return self.query(*args, **kwargs)
-
-    def query(self, *args, **kwargs):
-        """
-        Execute query
-
-        :param sql_text: SQL query text
-        :param args: SQL query values
-        :return:
-        :rtype: QueryResult
-        """
-        return self.connection._query(*args, **kwargs)
-
-    def select_field(self, table, what, where):
-        res = self.select(table, what, where)
-        if res.num_rows != 1:
-            raise ValueError("Query in select_field() returned {} rows instead of 1".format(res.num_rows))
-        return res.all_rows[0][0]
 
 
 class QueryResult(object):
